% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cbind.R
\name{c_cbind}
\alias{c_cbind}
\title{Bind columns of two c_mseries objects}
\usage{
c_cbind(cx1 = NULL, cx2 = NULL)
}
\arguments{
\item{cx1}{c_mseries; primary object}

\item{cx2}{c_mseries; object to append}
}
\value{
c_mseries object with data columns form c_mseries input objects
}
\description{
Bind new columns of c_mseries objects
}
