% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_extrapolate_dir_shear.R
\name{c_extrapolate_dir_shear}
\alias{c_extrapolate_dir_shear}
\title{Wind speed extrapolation with directional shear}
\usage{
c_extrapolate_dir_shear(
  cx = NULL,
  ws_signal = NULL,
  dir_signal = NULL,
  ws_min = NULL,
  dir_shear = NULL,
  height = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{ws_signal}{character; name of wind speed signal to extrapolate.}

\item{dir_signal}{character; name of direction signal to use.}

\item{ws_min}{numeric; minimal wind speed at the lower sensors.}

\item{dir_shear}{numeric vector with shear values for 16 sectors.}

\item{height}{numeric; height to extrapolate.}

\item{start_date}{character; start date in the format YYYY-MM-DD.}

\item{end_date}{character; end date in the format YYYY-MM-DD.}
}
\value{
Function returns an extrapolated c_mseries wind speed series
}
\description{
Create an extrapolated wind speed series with provided\cr
shear values for 16 sectors.
}
