% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_find_weibull.R
\name{c_find_weibull}
\alias{c_find_weibull}
\title{c_find_weibull}
\usage{
c_find_weibull(
  cx = NULL,
  signal = NULL,
  kmin = 1,
  kmax = 8,
  eps = 0.000001,
  iter = 50
)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{signal}{character, a name of a wind speed signal. If NULL, the main\cr
signal is used.}

\item{kmin}{numeric, minimum value of the shape k factor.}

\item{kmax}{numeric, maximum value of the shape k factor.}

\item{eps}{numeric, accuracy, the default value is 0.000001.}

\item{iter}{numeric, number of iterations.}
}
\value{
list with the scale c factor and the shape k factor\cr
of the Weibull distribution.
}
\description{
Find the scale c factor and the shape k factor of the Weibull distribution
for c_mseries, a selected wind speed signal and a given range of k.
}
\examples{
\dontrun{
c_find_weibull(cx)
c_find_weibull(cx, signal = "WS77")
c_find_weibull(cx, signal = "WS77", kmin = 1, kmax = 8, eps = 0.000001, iter = 50)
}

}
