% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_means.R
\name{c_means}
\alias{c_means}
\title{Monthly means or Means of Monthyl Means}
\usage{
c_means(cx = NULL, mom = FALSE, signals = NULL)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement.\cr
or a numeric vector.}

\item{mom}{logical, if TRUE, means of monthly means will be calculated.}

\item{signals}{character, name(s) of signal(s). If NULL, the main.\cr
signal is used.}
}
\value{
Function returns a tibble with monthly means or with means of monthly means.
}
\description{
Function create a table with monthly means for all or selected signals.
It is also possible to create a table with means of monthly means.
}
\examples{
\dontrun{
c_means(cx)
c_means(cx, mom = TRUE)
}

}
