% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_mseries.R
\name{c_mseries}
\alias{c_mseries}
\title{c_mseries}
\usage{
c_mseries(mdata, ...)
}
\arguments{
\item{mdata}{data frame; wind measurements data, names of columns are required.}

\item{date_col}{character; name of column with dates and hours.\cr
A selected column should be POSIXct.}

\item{ws_col}{character; vector with names of columns with wind speed signals.}

\item{ws_h}{numeric; vector with heights of wind speed signals.
#' @param ws_sd_col character; vector with names of columns with standard deviation of\cr
wind speed signals.}

\item{dir_col}{character; vector with names of columns with wind direction signals.}

\item{dir_h}{numeric; vector with heights of wind direction signals.}

\item{dir_sd_col}{character; vector with names of columns with standard deviation of\cr
wind direction signals}

\item{t_col}{character; vector with names of columns with temperature signals.}

\item{t_h}{numeric; vector with heights of temperature signals.}

\item{p_col}{character; vector with names of columns with pressure signals.}

\item{p_c}{numeric; vector with heights of pressure signals.}

\item{ws_main}{character; name of main wind speed signal. If NULL, the highest\cr
signal will be selected.}

\item{wd_main}{character; name of main wind direction signal. If NULL, the highest\cr
signal will be selected.}

\item{name}{character; name of measurement site.}

\item{tzone}{character; tzone = "UTC"}
}
\value{
c_mseries object with measurement data and information about a measurement site
}
\description{
Create c_mseries object with data and information about wind measurement.
The output c_mseries object consists of:
\itemize{
\item mdata in xts Time-Series Object  with wind measurement data
\item t_index with dates
\item start_date, a start date of measurement data
\item end_date, an end date of measurement data
\item wind_speed, names and heights of measured wind speed signals
\item wind_dir, names and heights of measured wind direction signals
\item wind_sd, names and heights of measured standard deviation
\item temp, names and heights of temperature measured signals
\item pressure, names and heights of winn speed measured signals
\item main_wind_speed, main wind speed measured signal
\item wind_dir, main wind direction measured signals
}
}
\examples{
\dontrun{
# Creating c_mseries object from the din data frame where
# the column DateTime consists of the time stamps in
the format 1999-01-01 00:10:00:
df <- c_mseries(mdata = din, date_col = "DateTime",
ws_col = c("WS125", "WS77", "WS44"), dir_col = c("WD77", "WD125"),
ws_h = c(125, 77, 44), dir_h = c(77, 125),
t_col = c("T3", "T44", "T118"), t_h = c(3, 44, 118),
p_col = "P8", p_h = 8,
name = "Test mast")
}

}
