% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_scale_offset.R
\name{c_scale_offset}
\alias{c_scale_offset}
\title{Scale and offset a signal}
\usage{
c_scale_offset(cx = NULL, signal = NULL, scale = NULL, offset = NULL)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{signal}{character, a name of a wind speed signal. If NULL, the main\cr
signal is used.}

\item{scale}{numeric, a value of a scale parameter}

\item{offset}{numeric, a value of a offset parameter}
}
\value{
c_mseries object with a provided signal multiply by scale and with\cr
added offset
}
\description{
Scale and offset a provided signal
}
\examples{
\dontrun{
}

}
