% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_shear.R
\name{c_shear}
\alias{c_shear}
\title{c_shear}
\usage{
c_shear(cx = NULL, ws_signals = NULL)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{ws_signals}{character, name(s) of wind speed signal(s). If NULL, the main\cr
signal is used.}
}
\description{
Create a c_shear object
}
