% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_signal_summary.R
\name{c_signal_summary}
\alias{c_signal_summary}
\title{c_signal_summary}
\usage{
c_signal_summary(
  cx = NULL,
  signal = NULL,
  start_date = NULL,
  end_date = NULL,
  numeric_months = TRUE
)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{signal}{character, name of signal. If NULL, the main\cr
signal is used.}

\item{start_date}{character; start date in the format YYYY-MM-DD.}

\item{end_date}{character; end date in the format YYYY-MM-DD.}

\item{numeric_months}{logical, if TRUE it returns months as numbers,\cr
otherwise as months names}
}
\value{
Function returns a tibble with a signal summary.
}
\description{
Function presents a summary of one signal
}
