% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_turbine.R
\name{c_turbine}
\alias{c_turbine}
\title{c_turbine}
\usage{
c_turbine(
  name,
  producer,
  model,
  power,
  diameter,
  hh,
  cut_in,
  cut_out,
  class_IEC,
  subclass_IEC,
  ws,
  p,
  cp,
  ct,
  ro
)
}
\arguments{
\item{name}{character; a name of wind turbine.}

\item{producer}{character; a producer name.}

\item{model}{character; a wind turbine model.}

\item{power}{character; a nominal power of wind turbine.}

\item{diameter}{numeric; a rotor diameter of wind turbine.}

\item{hh}{numeric; a hub height of wind turbine.}

\item{cut_in}{numeric; a cut-in wind speed of wind turbine}

\item{cut_out}{numeric; a cut-out wind speed of wind turbine}

\item{class_IEC}{character; a wind turbine IEC class}

\item{subclass_IEC}{character; a wind turbine IEC subclass}

\item{ws}{numeric; a vector with wind speed bins}

\item{p}{numeric; a vector with power for each wind speed bin}

\item{cp}{numeric; a vector with power coefficient for each wind speed bin}

\item{ct}{numeric; a vector with thrust coefficient for each wind speed bin}

\item{ro}{numeric: air density}
}
\value{
c_turbine object with information about a wind turbine
}
\description{
Create c_turbine object with basic information about a wind turbine.
\itemize{
\item name, a wind turbine name
\item producer
\item model, a wind turbine model
\item power, a nominal power of wind turbine in kW
\item diameter, a rotor diameter of wind turbine
\item hh, a hub height of wind turbine in meters
\item cut_in, a cut in wind speed in m/s
\item cut_out, a cut out wind speed in m/s
\item class_IEC, a wind turbine IEC class
\item subclass_IEC, a wind turbine subclass IEC
\item ws, wind speed bins, values in m/s
\item p, power in kW
\item cp, power coefficient
\item ct, thrust coefficient
\item ro, air density in kg/m3
}
}
\examples{
\dontrun{
}

}
