% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_weibull.R
\name{c_weibull}
\alias{c_weibull}
\title{Weibull distribution}
\usage{
c_weibull(c, k, ws = 5, ro = 1.225, ...)
}
\arguments{
\item{c}{scale factor}

\item{k}{shape factor}

\item{ws}{wind speed}

\item{ro}{air density}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Function return a list of numeric values of a probability for a given wind speed.
Additionally for the given c and k the function returns mean wind speed, median wind speed,
maximum probability, wind speed for maximum probability, power density, scale factor,
shape factor
}
\description{
Weibull distribution function calculates probability for a given wind speed, a
shape factor and a scale factor.
}
\examples{
c_weibull(c = 7.0, k = 2.0, ws = 5.0, ro = 1.225)
c_weibull(7, 2, 5, 1.225)

}
