% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_windrose.R
\name{c_windrose}
\alias{c_windrose}
\title{Windrose plot}
\usage{
c_windrose(cx, signals = NULL, col = "blue")
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{signals}{character, a name of a wind direction signal(s). If NULL, the main\cr
signal is used.}

\item{col}{color}
}
\description{
Plot a windrose from c_mseries object or a numeric vector.
}
