% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_mseries.R
\name{is_c_mseries}
\alias{is_c_mseries}
\title{Test if the object is a c_mseries}
\usage{
is_c_mseries(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `c_mseries` class.
}
\description{
This function returns `TRUE` for c_mseries,
and `FALSE` for all other objects.
}
