% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_turbine.R
\name{is_c_turbine}
\alias{is_c_turbine}
\title{Test if the object is a c_turbine}
\usage{
is_c_turbine(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `c_turbine` class.
}
\description{
This function returns `TRUE` for c_turbine,
and `FALSE` for all other objects.
}
