% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_weibull.R
\name{is_c_weibull}
\alias{is_c_weibull}
\title{Test if the object is a c_weibull object class}
\usage{
is_c_weibull(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `c_weibull` class.
}
\description{
This function returns `TRUE` for c_mseries,
and `FALSE` for all other objects.
}
