% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cbind.R
\name{c_cbind}
\alias{c_cbind}
\title{Bind columns of two c_mseries objects}
\usage{
c_cbind(cx1 = NULL, cx2 = NULL)
}
\arguments{
\item{cx1;}{first c_mseries object}

\item{cx2;}{second c_mseries object}
}
\description{
Bind new columns of c_mseries objects
}
