% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_coverage.R
\name{c_coverage}
\alias{c_coverage}
\title{c_coverage}
\usage{
c_coverage(cx = NULL, signals = NULL)
}
\description{
Function presents coverage of wind measuremetn data.
#' @param cx c_mseries object with data and information about wind measurement\cr
or a numeric vector.
#' @param signals character, name(s) of signal(s). If NULL, the main\cr
signal is used.
}
