% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_means.R
\name{c_means}
\alias{c_means}
\title{c_means}
\usage{
c_means(cx = NULL, mom = FALSE)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{mom}{logical, if TRUE, means of monthly means will be calculated}
}
\description{
c_means
}
