% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_weibull_plot.R
\name{c_weibull_plot}
\alias{c_weibull_plot}
\title{c_weibull_plot}
\usage{
c_weibull_plot(
  cx = NULL,
  signal = NULL,
  c = 7,
  k = 2,
  vmin = 0,
  vmax = 25,
  line = TRUE,
  lcolor = "blue",
  lsize = 1,
  hist = FALSE,
  hcolor = "#DD8888"
)
}
\arguments{
\item{cx}{c_mseries object with data and information about wind measurement\cr
or a numeric vector.}

\item{signal}{character, a name of a wind speed signal. If NULL, the main\cr
signal is used.}

\item{c}{numeric, scale factor}

\item{k}{numeric, shape factor}

\item{vmin}{numeric, minimal wind speed}

\item{vmax}{numeric, maximum wind speed}

\item{line}{logical, if TRUE, plot Weibull line}

\item{lcolor}{character, line color of Weibull distribution}

\item{lsize}{numeric, line size of Weibull distribution}

\item{hist}{logical, if TRUE, plot histogram}

\item{hcolor}{character, histogram color}
}
\description{
Plot Weibull distribution from c_mseries object or from a numeric vector\cr
or with given c scale factor and k shape factor
}
\examples{
\dontrun{
c_weibull_plot(cs)
c_weibull_plot(cs, hist = TRUE)
}
c_weibull_plot(c = 7.0, k = 2.0, vmin = 0, vmax = 25, lcolor = "blue")
c_weibull_plot(c = 7.0, k = 2.0, vmin = 0, vmax = 25, lcolor = "red", lsize = 1, hist = TRUE)
c_weibull_plot(c = 7.0, k = 2.0, vmin = 0, vmax = 30, lcolor = "red", hist = TRUE, hcolor = "blue")

}
