% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefiro-package.R
\docType{package}
\name{cefiro-package}
\alias{cefiro}
\alias{cefiro-package}
\title{cefiro: Simple Calculations and Plots for Wind Measurement Data}
\description{
Cefiro
}
\author{
\strong{Maintainer}: Pawel Cwiek \email{prc.altodato@gmail.com}

Other contributors:
\itemize{
  \item Kurt Schaldemose Hansen (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Resource data. Technical University of Denmark. Collection. https://doi.org/10.11583/DTU.c.5405286.v4) [contributor]
  \item Nikola Vasiljevic (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Resource data. Technical University of Denmark. Collection. https://doi.org/10.11583/DTU.c.5405286.v4) [contributor]
  \item Steen Arne Sørensen (Hansen, Kurt Schaldemose; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Resource data. Technical University of Denmark. Collection. https://doi.org/10.11583/DTU.c.5405286.v4) [contributor]
}

}
\keyword{internal}
