% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefiro-package.R
\docType{package}
\name{cefiro-package}
\alias{cefiro}
\alias{cefiro-package}
\title{cefiro: Simple calculations and plots for wind measurement data}
\description{
Cefiro
}
\author{
\strong{Maintainer}: Pawel Cwiek \email{prc.altodato@gmail.com}

Other contributors:
\itemize{
  \item Kurt Schaldemose Hansen (Hansen, ; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
  \item Nikola Vasiljevic (Hansen, ; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
  \item Steen Arne Sørensen (Hansen, ; Vasiljevic, Nikola; Sørensen, Steen Arne (2021). Wind resource data from the tall Risø met mast. Technical University of Denmark. Dataset. https://doi.org/10.11583/DTU.14153204.v1) [contributor]
}

}
\keyword{internal}
